requirements.txt
streamlit
pymongo
dnspython
paho-mqtt
requests